chore: lints & formatting

+++ testdata/zk/InvariantDeposit.t.sol
@@ -13,16 +13,19 @@ abstract contract StdInvariant {
     }
 
     address[] internal _targetedContracts;
+
     function targetContracts() public view returns (address[] memory) {
         return _targetedContracts;
     }
 
     FuzzSelector[] internal _targetedSelectors;
+
     function targetSelectors() public view returns (FuzzSelector[] memory) {
         return _targetedSelectors;
     }
 
     address[] internal _targetedSenders;
+
     function targetSenders() public view returns (address[] memory) {
         return _targetedSenders;
     }
@@ -32,7 +35,7 @@ contract ZkInvariantTest is DSTest, StdInvariant {
     Vm constant vm = Vm(HEVM_ADDRESS);
     Deposit deposit;
 
-    uint constant dealAmount = 1 ether;
+    uint256 constant dealAmount = 1 ether;
 
     function setUp() external {
         // to fund for fees
@@ -41,8 +44,8 @@ contract ZkInvariantTest is DSTest, StdInvariant {
         _targetedSenders.push(address(65536 + 123));
         _targetedSenders.push(address(65536 + 1234));
 
-        for (uint i = 0; i < _targetedSenders.length; i++) {
-          vm.deal(_targetedSenders[i], dealAmount); // to pay fees
+        for (uint256 i = 0; i < _targetedSenders.length; i++) {
+            vm.deal(_targetedSenders[i], dealAmount); // to pay fees
         }
 
         deposit = new Deposit();
